pyproject.toml
[project]
name = "postblog"
version = "0.1.0"
description = "高速ブログ投稿支援アプリ - インプットした知識を最速でブログ記事としてアウトプット"
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }
authors = [{ name = "Your Name", email = "your.email@example.com" }]
keywords = ["blog", "posting", "openai", "seo"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]

dependencies = [
    "customtkinter>=5.2.0",
    "openai>=1.0.0",
    "httpx>=0.27.0",
    "keyring>=25.0.0",
    "mistune>=3.0.0",
    "tomli-w>=1.0.0",
    "gitpython>=3.1.0",
    "Pillow>=10.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.23.0",
    "mypy>=1.8.0",
    "ruff>=0.4.0",
    "pre-commit>=3.6.0",
]

[project.scripts]
postblog = "postblog.app:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/postblog"]

# =============================================================================
# Ruff - Linter & Formatter (PEP 8準拠)
# =============================================================================
[tool.ruff]
target-version = "py312"
line-length = 88
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
    "TCH",    # flake8-type-checking
    "PTH",    # flake8-use-pathlib
    "ERA",    # eradicate (commented-out code)
    "PL",     # Pylint
    "RUF",    # Ruff-specific rules
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "PLR0913", # too many arguments
    "PLR2004", # magic value comparison
    "RUF001", # ambiguous Unicode character in string (日本語全角文字)
    "RUF002", # ambiguous Unicode character in docstring (日本語全角文字)
    "RUF003", # ambiguous Unicode character in comment (日本語全角文字)
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = [
    "ARG",    # Unused arguments are common in tests (fixtures)
    "PLR2004", # Magic values are acceptable in tests
]
"src/postblog/gui/**/*.py" = [
    "ARG002", # Unused method arguments are common in GUI event handlers
    "ARG005", # Unused lambda arguments are common in tkinter bind callbacks
    "PLW0108", # Lambda wrappers are needed for tkinter callbacks with parameters
    "PLR0915", # GUI build methods legitimately have many statements
    "TC003",  # Type-checking imports conflict with runtime use in GUI
    "ERA001", # Comments in Japanese are falsely detected as commented-out code
    "SIM105", # try-except-pass is acceptable for graceful error handling in GUI
    "PLC0415", # Import inside function is acceptable for contextlib in GUI
]

[tool.ruff.lint.isort]
known-first-party = ["postblog"]
force-single-line = false
lines-after-imports = 2

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
docstring-code-format = true

# =============================================================================
# MyPy - 型チェック
# =============================================================================
[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_any_generics = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_configs = true
show_error_codes = true
show_column_numbers = true

# サードパーティライブラリの型スタブがない場合の設定
[[tool.mypy.overrides]]
module = ["tests.*"]
disallow_untyped_defs = false

[[tool.mypy.overrides]]
module = [
    "customtkinter.*",
    "PIL.*",
    "keyring.*",
    "git.*",
    "mistune.*",
]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["postblog.gui.*"]
disable_error_code = ["misc", "no-any-return", "assignment"]

# =============================================================================
# Pytest - テスト
# =============================================================================
[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--strict-markers",
    "--tb=short",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
]
asyncio_mode = "auto"

# =============================================================================
# Coverage - カバレッジ
# =============================================================================
[tool.coverage.run]
source = ["src/postblog"]
branch = true
omit = ["*/tests/*", "*/__pycache__/*", "src/postblog/gui/*", "src/postblog/app.py"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]
fail_under = 80
show_missing = true
